+++ SECURITY.md
@@ -3,6 +3,5 @@
 ## Supported Versions
 | Version | Supported          |
 | ------- | ------------------ |
-| 0.2.2   | :white_check_mark: |
-| 0.2.1   | :x:                |
+| 0.2.5   | :white_check_mark: |
 | 0.1.5   | :x:                |
