+++ rvas-core_v0.2/pom.xml
@@ -69,7 +69,7 @@
     </dependencies>
 
     <build>
-        <finalName>rvas-core-0.2.5-258</finalName>
+        <finalName>rvas-core-0.2.5-259</finalName>
         <plugins>
             <!-- download source code in Eclipse, best practice-->
             <plugin>
