+++ rvas-core_v0.2/src/main/java/core/backend/ChatPrint.java
@@ -23,7 +23,10 @@ package core.backend;
  * 
  * */
 
-import core.objects.*;
+import core.data.PVPdata;
+import core.data.PlayerMeta;
+import core.data.objects.*;
+import core.backend.utils.Util;
 
 import java.util.*;
 import java.text.SimpleDateFormat;
@@ -89,7 +92,7 @@ public class ChatPrint {
 			
 			if (target_name == null) {
 				
-				TextComponent b = new TextComponent("[unknown], " + Utilities.timeToString(realLeaders_0_15.get(pid)));
+				TextComponent b = new TextComponent("[unknown], " + Util.timeToString(realLeaders_0_15.get(pid)));
 				TextComponent c = new TextComponent(a1, b);
 				
 				c.setColor(ChatColor.GOLD);
@@ -103,7 +106,7 @@ public class ChatPrint {
 				String kd = PVPdata.getStats(offPlayer).kd;
 				
 				TextComponent a2 = new TextComponent(target_name + ", ");
-				TextComponent b = new TextComponent(Utilities.timeToString(realLeaders_0_15.get(pid)));
+				TextComponent b = new TextComponent(Util.timeToString(realLeaders_0_15.get(pid)));
 				
 				HoverEvent hoverStats = new HoverEvent(HoverEvent.Action.SHOW_TEXT, new Text("Kills: "+kills+" | Deaths: "+deaths+" | K/D: "+kd));
 				ClickEvent shortcut = new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/stats " + target_name);
@@ -169,7 +172,7 @@ public class ChatPrint {
 		TextComponent rank_a = new TextComponent("Ranking: ");
 		TextComponent rank_b = new TextComponent("" + PlayerMeta.getRank(target));
 		TextComponent playtime_a = new TextComponent("Time played: ");
-		TextComponent playtime_b = new TextComponent(Utilities.timeToString(PlayerMeta.getPlaytime(target)));
+		TextComponent playtime_b = new TextComponent(Util.timeToString(PlayerMeta.getPlaytime(target)));
 
 		double hours = PlayerMeta.getPlaytime(target) / 3600;
 		Text playtime_hover = new Text(new DecimalFormat("0.00").format(hours) + " hours");
