+++ rvas-core_v0.2/src/main/java/core/backend/ItemCheck.java
@@ -2,6 +2,7 @@ package core.backend;
 
 import java.util.*;
 
+import core.data.PlayerMeta;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.block.ShulkerBox;
