+++ rvas-core_v0.2/src/main/java/core/commands/Discord.java
@@ -23,7 +23,7 @@ package core.commands;
  * */
 
 import core.backend.Config;
-import core.backend.PlayerMeta;
+import core.data.PlayerMeta;
 import core.tasks.Analytics;
 
 import org.bukkit.entity.Player;
