+++ rvas-core_v0.2/src/main/java/core/commands/Last.java
@@ -22,7 +22,7 @@ package core.commands;
  *
  * */
 
-import core.backend.Reversed;
+import core.backend.utils.Reversed;
 
 import java.util.UUID;
 import org.bukkit.entity.Player;
