+++ rvas-core_v0.2/src/main/java/core/commands/Message.java
@@ -1,7 +1,7 @@
 package core.commands;
 
 import core.tasks.Analytics;
-import core.backend.PlayerMeta;
+import core.data.PlayerMeta;
 
 import java.util.*;
