+++ rvas-core_v0.2/src/main/java/core/commands/Reply.java
@@ -1,6 +1,6 @@
 package core.commands;
 
-import core.backend.PlayerMeta;
+import core.data.PlayerMeta;
 import core.tasks.Analytics;
 
 import java.util.Arrays;
