+++ rvas-core_v0.2/src/main/java/core/data/Aliases.java
@@ -1,4 +1,4 @@
-package core.backend;
+package core.data;
 
 /* *
  *
