+++ rvas-core_v0.2/src/main/java/core/data/objects/PlayerSettings.java
@@ -1,4 +1,4 @@
-package core.objects;
+package core.data.objects;
 
 /* *
  *
