+++ rvas-core_v0.2/src/main/java/core/events/MoveListener.java
@@ -23,26 +23,21 @@ package core.events;
  * */
 
 import core.backend.Config;
-import core.backend.PlayerMeta;
-import core.backend.Utilities;
-import core.commands.AFK;
+import core.data.PlayerMeta;
+import core.backend.utils.Util;
 
 import java.util.*;
-import net.md_5.bungee.api.ChatColor;
-import net.md_5.bungee.api.chat.TextComponent;
-
 import org.bukkit.GameMode;
 import org.bukkit.Location;
-import org.bukkit.World;
 
 import org.bukkit.World.Environment;
 import org.bukkit.entity.EnderCrystal;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
 
+import org.bukkit.event.Listener;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntityPortalEvent;
 import org.bukkit.event.player.PlayerMoveEvent;
 
@@ -114,7 +109,7 @@ public class MoveListener implements Listener {
 				if (portalFrom.equals(Environment.THE_END)) {
 					e.setCancelled(true);
 
-					World overworld = Utilities.getWorldByDimension(Environment.NORMAL);
+					org.bukkit.World overworld = Util.getWorldByDimension(Environment.NORMAL);
 					if (overworld == null) {
 						System.out.println("WARN couldn't find NORMAL dimension onEntityPortal()");
 						return;
